requirements.txt
torch
torchvision
streamlit
scikit-learn
matplotlib
seaborn
tqdm
pillow
numpy